Rename include dir to swingby.
