+++ examples/00_application/main.c
@@ -1,4 +1,4 @@
-#include <foundation/application.h>
+#include <swingby/application.h>
 
 int main(int argc, char *argv[])
 {
