+++ examples/02_desktop-surface/main.c
@@ -1,4 +1,4 @@
-#include <foundation/foundation.h>
+#include <swingby/swingby.h>
 
 int main(int argc, char *argv[])
 {
