--- include/foundation/foundation.h
@@ -1,19 +1,0 @@
-#ifndef _FOUNDATION_FOUNDATION_H
-#define _FOUNDATION_FOUNDATION_H
-
-#include <foundation/application.h>
-#include <foundation/color.h>
-#include <foundation/desktop-surface.h>
-#include <foundation/egl-context.h>
-#include <foundation/event-dispatcher.h>
-#include <foundation/event.h>
-#include <foundation/image.h>
-#include <foundation/input.h>
-#include <foundation/log.h>
-#include <foundation/point.h>
-#include <foundation/rect.h>
-#include <foundation/size.h>
-#include <foundation/surface.h>
-#include <foundation/view.h>
-
-#endif /* _FOUNDATION_FOUNDATION_H */
