+++ include/swingby/desktop-surface.h
@@ -1,7 +1,7 @@
 #ifndef _FOUNDATION_DESKTOP_SURFACE_H
 #define _FOUNDATION_DESKTOP_SURFACE_H
 
-#include <foundation/event.h>
+#include <swingby/event.h>
 
 typedef struct ft_size_i_t ft_size_i_t;
 typedef struct ft_rect_i_t ft_rect_i_t;
