+++ include/swingby/event.h
@@ -3,9 +3,9 @@
 
 #include <stdbool.h>
 
-#include <foundation/input.h>
-#include <foundation/point.h>
-#include <foundation/size.h>
+#include <swingby/input.h>
+#include <swingby/point.h>
+#include <swingby/size.h>
 
 #ifdef __cplusplus
 extern "C" {
