+++ include/swingby/swingby.h
@@ -1,0 +1,19 @@
+#ifndef _FOUNDATION_FOUNDATION_H
+#define _FOUNDATION_FOUNDATION_H
+
+#include <swingby/application.h>
+#include <swingby/color.h>
+#include <swingby/desktop-surface.h>
+#include <swingby/egl-context.h>
+#include <swingby/event-dispatcher.h>
+#include <swingby/event.h>
+#include <swingby/image.h>
+#include <swingby/input.h>
+#include <swingby/log.h>
+#include <swingby/point.h>
+#include <swingby/rect.h>
+#include <swingby/size.h>
+#include <swingby/surface.h>
+#include <swingby/view.h>
+
+#endif /* _FOUNDATION_FOUNDATION_H */
