+++ src/bench.c
@@ -1,4 +1,4 @@
-#include <foundation/bench.h>
+#include <swingby/bench.h>
 
 #include <stdlib.h>
 #include <stdio.h>
