+++ src/image.c
@@ -1,9 +1,9 @@
-#include <foundation/image.h>
+#include <swingby/image.h>
 
 #include <stdint.h>
 #include <stdlib.h>
 
-#include <foundation/log.h>
+#include <swingby/log.h>
 
 #ifdef __cplusplus
 extern "C" {
