+++ src/output.c
@@ -1,4 +1,4 @@
-#include <foundation/output.h>
+#include <swingby/output.h>
 
 #include <stdlib.h>
