+++ src/rect.c
@@ -1,4 +1,4 @@
-#include <foundation/rect.h>
+#include <swingby/rect.h>
 
 #ifdef __cplusplus
 extern "C" {
