+++ src/size.c
@@ -1,4 +1,4 @@
-#include <foundation/size.h>
+#include <swingby/size.h>
 
 #ifdef __cplusplus
 extern "C" {
